+++ framework/CHANGELOG.md
@@ -382,6 +382,7 @@ Yii Framework 2 Change Log
 2.0.7 February 14, 2016
 -----------------------
 
+- Bug #6526: Fixed `yii\db\ColumnSchema::typecast()` returning value with comma as decimal mark if corresponding locale set. (leammas)
 - Bug #6351: Find MySQL FK constraints from `information_schema` tables instead of `SHOW CREATE TABLE` to improve reliability (nineinchnick)
 - Bug #6363, #8301, #8582, #9566: Fixed data methods and PJAX issues when used together (derekisbusy)
 - Bug #6876: Fixed RBAC migration MSSQL cascade problem (thejahweh)